added updateTreeColor to utils from treeView
