+++ src/js/bookmarkManager/treeView.js
@@ -1,12 +1,16 @@
 import { section } from '../lib/react-clone';
 import File from '../components/File';
 import Folder from '../components/Folder';
+import options from '../config/config';
+import { updateTreeColor } from '../utils/treeView';
 import {
   createTree,
   handleFileClick,
   handleFolderHeaderClick
 } from './treeViewComponents';
-import options from '../config/config';
+
+// depends: displayFolderInfo, displayBookmark
+
 const {
   defaultFileIconColor,
   defaultFolderIconColor,
@@ -15,7 +19,6 @@ const {
 } = options;
 
 document.addEventListener('DOMContentLoaded', () => {
-  console.log(createTree, section);
   var sidenavs = document.querySelectorAll('.sidenav');
   M.Sidenav.init(sidenavs);
 
@@ -90,65 +93,3 @@ document.addEventListener('DOMContentLoaded', () => {
     setTimeout(updateTreeColor, 100);
   });
 });
-
-/**
- * Removes `oldColor` from `elem.classList` and adds `newColor`
- * @param {HTMLElement} elem
- * @param {string} oldColor
- * @param {string} newColor
- */
-function colorElement(elem, oldColor, newColor) {
-  elem.classList.remove(oldColor);
-  elem.classList.add(newColor);
-}
-
-function updateTreeColor(color = trackedFileIconColor) {
-  chrome.storage.sync.get(['dynBookmarks'], ({ dynBookmarks }) => {
-    let dynBook = dynBookmarks || {};
-    chrome.bookmarks.getTree((results) => {
-      const rootNode = results[0];
-      (function traverseTree(node) {
-        if (!node.children) {
-          const file = document
-            .getElementById(node.id)
-            .querySelector('.file-icon');
-          if (dynBook[node.id]) {
-            colorElement(file, defaultFileIconColor, trackedFileIconColor);
-            return true;
-          } else {
-            colorElement(file, trackedFileIconColor, defaultFileIconColor);
-            return false;
-          }
-        } else {
-          let hasTrackedChild = false;
-          for (let child of node.children) {
-            if (traverseTree(child)) {
-              hasTrackedChild = true;
-            }
-          }
-          const folder = document.getElementById(node.id);
-          if (!folder) {
-            // this can happen if the folder is root
-            return hasTrackedChild;
-          }
-          const folderIcon = folder.querySelector(
-            '.folder-header > .folder-icon'
-          );
-          if (hasTrackedChild) {
-            colorElement(
-              folderIcon,
-              defaultFolderIconColor,
-              trackedFolderIconColor
-            );
-          } else {
-            colorElement(
-              folderIcon,
-              trackedFolderIconColor,
-              defaultFolderIconColor
-            );
-          }
-        }
-      })(rootNode);
-    });
-  });
-}
